[LVL-3]: Refactor TopSecret to use entities

+++ src/main/java/com/quasar/model/SatelliteStatus.java
@@ -3,10 +3,10 @@ package com.quasar.model;
 public class SatelliteStatus {
 
     public String name;
-    public Float distance;
+    public Double distance;
     public String[] message;
 
-    public SatelliteStatus(String name, Float distance, String[] message) {
+    public SatelliteStatus(String name, Double distance, String[] message) {
         this.name = name;
         this.distance = distance;
         this.message = message;
